+++ client/src/pages/LogInPage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Typography } from '@mui/material';
+
+const LogInPage: React.FC = () => {
+  return (
+    <div>
+      <Typography variant="h2">Log In Page</Typography>
+    </div>
+  );
+};
+
+export default LogInPage;
