+++ client/src/pages/SignUpPage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Typography } from '@mui/material';
+
+const SignUpPage: React.FC = () => {
+  return (
+    <div>
+      <Typography variant="h2">Sign Up Page</Typography>
+    </div>
+  );
+};
+
+export default SignUpPage;
